backend/requirements.txt
# python version 3.11.9

annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.128.0
h11==0.16.0
idna==3.11
motor==3.7.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
python-multipart==0.0.22
bcrypt=4.0.1
ecdsa==0.19.1
limits==5.8.0
packaging==26.0
passlib==1.7.4
python-jose==3.5.0
six==1.17.0
slowapi==0.1.9
wrapt==2.1.1   

# ---- External APIs / stats ----
requests==2.32.5
PyGithub==2.8.1
google-genai=1.62.0
